refactor: re-organize views
