+++ frontend/app/views/common/warning_dialog.py
