+++ frontend/app/views/visitor/visitor_details_dialog.py
